Use long instand of sizeof to allow -1.

+++ fifo.c
@@ -116,10 +116,10 @@ void fifo_put(fifo_t * fifo, handle_t * handle, void * data)
 {
   node_t * node = _hzdptr_setv(&handle->enq, &handle->hazard);
 
-  size_t i  = fetch_and_add(&fifo->enq, 1);
-  size_t s  = fifo->size;
-  size_t ni = i / s;
-  size_t li = i % s;
+  long i  = fetch_and_add(&fifo->enq, 1);
+  long s  = fifo->size;
+  long ni = i / s;
+  long li = i % s;
 
   if (node->id != ni) {
     node = handle->enq = update(node, ni, s, &handle->winner);
@@ -134,10 +134,10 @@ void * fifo_get(fifo_t * fifo, handle_t * handle)
 {
   node_t * node = _hzdptr_setv(&handle->deq, &handle->hazard);
 
-  size_t i  = fetch_and_add(&fifo->deq, 1);
-  size_t s  = fifo->size;
-  size_t ni = i / s;
-  size_t li = i % s;
+  long i  = fetch_and_add(&fifo->deq, 1);
+  long s  = fifo->size;
+  long ni = i / s;
+  long li = i % s;
 
   if (node->id != ni) {
     node = handle->deq = update(node, ni, s, &handle->winner);
